+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
+## 0.3.0
+
+- Add hybrid flow
+- Add inapp browser client
+- Add logout support
+- Keep only flutter library - drop command line tool, web and io clients
+
 ## 0.2.4
 
-- Allow only signing algorithms specified in `id_token_signing_alg_values_supported` 
-parameter of issuer metadata
+- Allow only signing algorithms specified in `id_token_signing_alg_values_supported` parameter of issuer metadata
 
 ## 0.2.1
 
@@ -12,7 +18,6 @@ parameter of issuer metadata
 ## 0.2.0
 
 - Dart 2/flutter compatibility
-
 
 ## 0.1.0